Pull all data structures to data manager
