+++ App.tsx
@@ -1,11 +1,14 @@
 import React from 'react'
 import AppContainer from './src/components/app-container'
 import Navigator from './src/'
+import { DataManagerProvider } from './src/utils/data-manager'
 
 export default function App() {
   return (
     <AppContainer>
-      <Navigator />
+      <DataManagerProvider>
+        <Navigator />
+      </DataManagerProvider>
     </AppContainer>
   )
 }
